+++ mirrord/agent/src/main.rs
@@ -15,10 +15,6 @@ use std::{
 };
 
 use actix_codec::Framed;
-use cli::parse_args;
-use dns::DnsApi;
-use error::{AgentError, Result};
-use file::FileManager;
 use futures::{
     stream::{FuturesUnordered, StreamExt},
     SinkExt, TryFutureExt,
@@ -27,9 +23,6 @@ use mirrord_protocol::{
     pause::DaemonPauseTarget, ClientMessage, DaemonCodec, DaemonMessage, GetEnvVarsRequest,
     LogMessage,
 };
-use outgoing::{udp::UdpOutgoingApi, TcpOutgoingApi};
-use sniffer::{SnifferCommand, TcpConnectionSniffer, TcpSnifferApi};
-use steal::api::TcpStealerApi;
 use tokio::{
     net::{TcpListener, TcpStream},
     select,
@@ -44,11 +37,17 @@ use tracing_subscriber::{fmt::format::FmtSpan, prelude::*};
 use crate::{
     cli::Args,
     container_handle::ContainerHandle,
+    dns::DnsApi,
+    error::{AgentError, Result},
+    file::FileManager,
+    outgoing::{udp::UdpOutgoingApi, TcpOutgoingApi},
     runtime::get_container,
+    sniffer::{SnifferCommand, TcpConnectionSniffer, TcpSnifferApi},
     steal::{
+        api::TcpStealerApi,
         connection::TcpConnectionStealer,
         ip_tables::{
-            SafeIpTables, IPTABLE_MESH, IPTABLE_MESH_ENV, IPTABLE_PREROUTING,
+            IPTablesWrapper, SafeIpTables, IPTABLE_MESH, IPTABLE_MESH_ENV, IPTABLE_PREROUTING,
             IPTABLE_PREROUTING_ENV, IPTABLE_STANDARD, IPTABLE_STANDARD_ENV,
         },
         StealerCommand,
@@ -89,39 +88,35 @@ struct State {
 impl State {
     /// Return [`Err`] if container runtime operations failed.
     pub async fn new(args: &Args) -> Result<State> {
-        let container =
-            get_container(args.container_id.as_ref(), args.container_runtime.as_ref()).await?;
-
-        let container = match container {
-            Some(container) => Some(ContainerHandle::new(container).await?),
-            None => None,
-        };
-
-        // If we are in an ephemeral container, we use pid 1.
-        // if not, we use the pid of the target container or fallback to self
-        let pid = {
-            if args.ephemeral_container {
-                "1".to_string()
-            } else {
-                container
-                    .as_ref()
-                    .map(|h| h.pid().to_string())
-                    .unwrap_or_else(|| "self".to_string())
-            }
-        };
-
         let mut env: HashMap<String, String> = HashMap::new();
 
+        let (ephemeral, container, pid) = match &args.mode {
+            cli::Mode::Targeted {
+                container_id,
+                container_runtime,
+            } => {
+                let container =
+                    get_container(container_id.clone(), Some(container_runtime)).await?;
+
+                let container_handle = ContainerHandle::new(container).await?;
+                let pid = container_handle.pid().to_string();
+
+                env.extend(container_handle.raw_env().clone());
+
+                (false, Some(container_handle), pid)
+            }
+            cli::Mode::Ephemeral => {
+                // in ephemeral container, we get same env as the target container, so copy our env.
+                env.extend(std::env::vars());
+
+                // If we are in an ephemeral container, we use pid 1.
+                (true, None, "1".to_string())
+            }
+            // if not, we use the pid of the target container or fallback to self
+            cli::Mode::Targetless | cli::Mode::BlackboxTest => (false, None, "self".to_string()),
+        };
+
         let environ_path = PathBuf::from("/proc").join(pid).join("environ");
-
-        if let Some(container) = container.as_ref() {
-            env.extend(container.raw_env().clone());
-        }
-
-        // in ephemeral container, we get same env as the target container, so copy our env.
-        if args.ephemeral_container {
-            env.extend(std::env::vars())
-        }
 
         match env::get_proc_environ(environ_path).await {
             Ok(environ) => env.extend(environ.into_iter()),
@@ -135,7 +130,7 @@ impl State {
             index_allocator: Default::default(),
             container,
             env,
-            ephemeral: args.ephemeral_container,
+            ephemeral,
         })
     }
 
@@ -225,13 +220,27 @@ impl State {
     }
 }
 
+enum BackgroundTask<Command> {
+    Running(TaskStatus, Sender<Command>),
+    Disabled,
+}
+
+impl<Command> Clone for BackgroundTask<Command> {
+    fn clone(&self) -> Self {
+        match self {
+            BackgroundTask::Disabled => BackgroundTask::Disabled,
+            BackgroundTask::Running(status, sender) => {
+                BackgroundTask::Running(status.clone(), sender.clone())
+            }
+        }
+    }
+}
+
 /// Handles to background tasks used by [`ClientConnectionHandler`].
 #[derive(Clone)]
 struct BackgroundTasks {
-    sniffer_status: TaskStatus,
-    sniffer_sender: Sender<SnifferCommand>,
-    stealer_status: TaskStatus,
-    stealer_sender: Sender<StealerCommand>,
+    sniffer: BackgroundTask<SnifferCommand>,
+    stealer: BackgroundTask<StealerCommand>,
     dns_api: DnsApi,
 }
 
@@ -241,7 +250,7 @@ struct ClientConnectionHandler {
     file_manager: FileManager,
     stream: Framed<TcpStream, DaemonCodec>,
     tcp_sniffer_api: Option<TcpSnifferApi>,
-    tcp_stealer_api: TcpStealerApi,
+    tcp_stealer_api: Option<TcpStealerApi>,
     tcp_outgoing_api: TcpOutgoingApi,
     udp_outgoing_api: UdpOutgoingApi,
     dns_api: DnsApi,
@@ -270,48 +279,9 @@ impl ClientConnectionHandler {
 
         let file_manager = FileManager::new(pid.or_else(|| ephemeral.then_some(1)));
 
-        let tcp_sniffer_api = match TcpSnifferApi::new(
-            id,
-            bg_tasks.sniffer_sender,
-            bg_tasks.sniffer_status,
-            CHANNEL_SIZE,
-        )
-        .await
-        {
-            Ok(api) => Some(api),
-            Err(e) => {
-                let message = format!(
-                    "Failed to create TcpSnifferApi: {e}, this could be due to kernel version."
-                );
-                warn!(message);
-                let _ = stream
-                    .send(DaemonMessage::LogMessage(LogMessage::warn(message)))
-                    .await; // Ignore message send error.
-
-                None
-            }
-        };
-
-        let tcp_stealer_api = match TcpStealerApi::new(
-            id,
-            bg_tasks.stealer_sender,
-            bg_tasks.stealer_status,
-            CHANNEL_SIZE,
-            protocol_version,
-        )
-        .await
-        {
-            Ok(api) => api,
-            Err(e) => {
-                let _ = stream
-                    .send(DaemonMessage::Close(format!(
-                        "Failed to create TcpStealerApi: {e}."
-                    )))
-                    .await; // Ignore message send error.
-
-                Err(e)?
-            }
-        };
+        let tcp_sniffer_api = Self::ceate_sniffer_api(id, bg_tasks.sniffer, &mut stream).await;
+        let tcp_stealer_api =
+            Self::ceate_stealer_api(id, bg_tasks.stealer, protocol_version, &mut stream).await?;
 
         let tcp_outgoing_api = TcpOutgoingApi::new(pid);
         let udp_outgoing_api = UdpOutgoingApi::new(pid);
@@ -331,6 +301,66 @@ impl ClientConnectionHandler {
         };
 
         Ok(client_handler)
+    }
+
+    async fn ceate_sniffer_api(
+        id: ClientId,
+        task: BackgroundTask<SnifferCommand>,
+        stream: &mut Framed<TcpStream, DaemonCodec>,
+    ) -> Option<TcpSnifferApi> {
+        if let BackgroundTask::Running(sniffer_status, sniffer_sender) = task {
+            match TcpSnifferApi::new(id, sniffer_sender, sniffer_status, CHANNEL_SIZE).await {
+                Ok(api) => Some(api),
+                Err(e) => {
+                    let message = format!(
+                        "Failed to create TcpSnifferApi: {e}, this could be due to kernel version."
+                    );
+
+                    warn!(message);
+
+                    // Ignore message send error.
+                    let _ = stream
+                        .send(DaemonMessage::LogMessage(LogMessage::warn(message)))
+                        .await;
+
+                    None
+                }
+            }
+        } else {
+            None
+        }
+    }
+
+    async fn ceate_stealer_api(
+        id: ClientId,
+        task: BackgroundTask<StealerCommand>,
+        protocol_version: semver::Version,
+        stream: &mut Framed<TcpStream, DaemonCodec>,
+    ) -> Result<Option<TcpStealerApi>> {
+        if let BackgroundTask::Running(stealer_status, stealer_sender) = task {
+            match TcpStealerApi::new(
+                id,
+                stealer_sender,
+                stealer_status,
+                CHANNEL_SIZE,
+                protocol_version,
+            )
+            .await
+            {
+                Ok(api) => Ok(Some(api)),
+                Err(e) => {
+                    let _ = stream
+                        .send(DaemonMessage::Close(format!(
+                            "Failed to create TcpStealerApi: {e}."
+                        )))
+                        .await; // Ignore message send error.
+
+                    Err(e)?
+                }
+            }
+        } else {
+            Ok(None)
+        }
     }
 
     /// Starts a loop that handles client connection and state.
@@ -368,7 +398,13 @@ impl ClientConnectionHandler {
                     Ok(message) => self.respond(DaemonMessage::Tcp(message)).await?,
                     Err(e) => break e,
                 },
-                message = self.tcp_stealer_api.recv() => match message {
+                message = async {
+                    if let Some(ref mut stealer_api) = self.tcp_stealer_api {
+                        stealer_api.recv().await
+                    } else {
+                        unreachable!()
+                    }
+                }, if self.tcp_stealer_api.is_some() => match message {
                     Ok(message) => self.respond(DaemonMessage::TcpSteal(message)).await?,
                     Err(e) => break e,
                 },
@@ -453,7 +489,12 @@ impl ClientConnectionHandler {
                 }
             }
             ClientMessage::TcpSteal(message) => {
-                self.tcp_stealer_api.handle_client_message(message).await?
+                if let Some(tcp_stealer_api) = self.tcp_stealer_api.as_mut() {
+                    tcp_stealer_api.handle_client_message(message).await?
+                } else {
+                    warn!("received tcp steal request while not available");
+                    Err(AgentError::SnifferApiError)?
+                }
             }
             ClientMessage::Close => {
                 return Ok(false);
@@ -479,9 +520,11 @@ impl ClientConnectionHandler {
                 .await?;
             }
             ClientMessage::SwitchProtocolVersion(version) => {
-                self.tcp_stealer_api
-                    .switch_protocol_version(version.clone())
-                    .await?;
+                if let Some(tcp_stealer_api) = self.tcp_stealer_api.as_mut() {
+                    tcp_stealer_api
+                        .switch_protocol_version(version.clone())
+                        .await?;
+                }
 
                 self.respond(DaemonMessage::SwitchProtocolVersionResponse(version))
                     .await?;
@@ -495,8 +538,7 @@ impl ClientConnectionHandler {
 
 /// Initializes the agent's [`State`], channels, threads, and runs [`ClientConnectionHandler`]s.
 #[tracing::instrument(level = "trace")]
-async fn start_agent() -> Result<()> {
-    let args = parse_args();
+async fn start_agent(args: Args) -> Result<()> {
     trace!("Starting agent with args: {args:?}");
 
     let listener = TcpListener::bind(SocketAddrV4::new(
@@ -516,7 +558,9 @@ async fn start_agent() -> Result<()> {
 
     let dns_api = DnsApi::new(state.container_pid(), 1000);
 
-    let (sniffer_task, sniffer_status) = {
+    let (sniffer_task, sniffer_status) = if args.mode.is_targetless() {
+        (None, None)
+    } else {
         let cancellation_token = cancellation_token.clone();
         let watched_task = WatchedTask::new(
             TcpConnectionSniffer::TASK_NAME,
@@ -538,10 +582,12 @@ async fn start_agent() -> Result<()> {
             "net",
         );
 
-        (task, status)
+        (Some(task), Some(status))
     };
 
-    let (stealer_task, stealer_status) = {
+    let (stealer_task, stealer_status) = if args.mode.is_targetless() {
+        (None, None)
+    } else {
         let cancellation_token = cancellation_token.clone();
         let watched_task = WatchedTask::new(
             TcpConnectionStealer::TASK_NAME,
@@ -561,14 +607,16 @@ async fn start_agent() -> Result<()> {
             "net",
         );
 
-        (task, status)
+        (Some(task), Some(status))
     };
 
-    let mut bg_tasks = BackgroundTasks {
-        sniffer_status,
-        sniffer_sender: sniffer_command_tx,
-        stealer_status,
-        stealer_sender: stealer_command_tx,
+    let bg_tasks = BackgroundTasks {
+        sniffer: sniffer_status
+            .map(|status| BackgroundTask::Running(status, sniffer_command_tx))
+            .unwrap_or(BackgroundTask::Disabled),
+        stealer: stealer_status
+            .map(|status| BackgroundTask::Running(status, stealer_command_tx))
+            .unwrap_or(BackgroundTask::Disabled),
         dns_api,
     };
 
@@ -644,14 +692,26 @@ async fn start_agent() -> Result<()> {
     trace!("Agent shutting down.");
     drop(cancel_guard);
 
-    sniffer_task.join().map_err(|_| AgentError::JoinTask)?;
-    if let Some(err) = bg_tasks.sniffer_status.err().await {
-        error!("start_agent -> sniffer task failed with error: {}", err);
+    let BackgroundTasks {
+        sniffer, stealer, ..
+    } = bg_tasks;
+
+    if let (Some(sniffer_task), BackgroundTask::Running(mut sniffer_status, _)) =
+        (sniffer_task, sniffer)
+    {
+        sniffer_task.join().map_err(|_| AgentError::JoinTask)?;
+        if let Some(err) = sniffer_status.err().await {
+            error!("start_agent -> sniffer task failed with error: {}", err);
+        }
     }
 
-    stealer_task.join().map_err(|_| AgentError::JoinTask)?;
-    if let Some(err) = bg_tasks.stealer_status.err().await {
-        error!("start_agent -> stealer task failed with error: {}", err);
+    if let (Some(stealer_task), BackgroundTask::Running(mut stealer_status, _)) =
+        (stealer_task, stealer)
+    {
+        stealer_task.join().map_err(|_| AgentError::JoinTask)?;
+        if let Some(err) = stealer_status.err().await {
+            error!("start_agent -> stealer task failed with error: {}", err);
+        }
     }
 
     trace!("Agent shutdown.");
@@ -661,7 +721,10 @@ async fn start_agent() -> Result<()> {
 async fn clear_iptable_chain() -> Result<()> {
     let ipt = iptables::new(false).unwrap();
 
-    SafeIpTables::load(ipt, false).await?.cleanup().await?;
+    SafeIpTables::load(IPTablesWrapper::from(ipt), false)
+        .await?
+        .cleanup()
+        .await?;
 
     Ok(())
 }
@@ -679,10 +742,9 @@ fn spawn_child_agent() -> Result<()> {
     Ok(())
 }
 
-async fn start_iptable_guard() -> Result<()> {
+async fn start_iptable_guard(args: Args) -> Result<()> {
     debug!("start_iptable_guard -> Initializing iptable-guard.");
 
-    let args = parse_args();
     let state = State::new(&args).await?;
     let pid = state.container_pid();
 
@@ -704,7 +766,7 @@ async fn start_iptable_guard() -> Result<()> {
     result
 }
 
-#[tokio::main]
+#[tokio::main(flavor = "current_thread")]
 async fn main() -> Result<()> {
     tracing_subscriber::registry()
         .with(
@@ -721,13 +783,19 @@ async fn main() -> Result<()> {
         env!("CARGO_PKG_VERSION")
     );
 
-    let agent_result = if std::env::var(IPTABLE_PREROUTING_ENV).is_ok()
-        && std::env::var(IPTABLE_MESH_ENV).is_ok()
+    let args = cli::parse_args();
+
+    let agent_result = if args.mode.is_targetless()
+        || (std::env::var(IPTABLE_PREROUTING_ENV).is_ok()
+            && std::env::var(IPTABLE_MESH_ENV).is_ok())
     {
-        start_agent().await
+        start_agent(args).await
     } else {
-        start_iptable_guard().await
+        start_iptable_guard(args).await
     };
+
+    // wait for background tasks to finish
+    tokio::time::sleep(std::time::Duration::from_secs(1)).await;
 
     match agent_result {
         Ok(_) => {